+++ backend/controllers/shoeController.js
@@ -73,15 +73,20 @@ exports.updateModels = async (req, res, next) => {
 exports.deleteModel = async (req, res, next) => {
     try {
         const id=req.originalUrl.split('/')[2]
-        await client.connect();
-        const result = await client.db('inventory').collection('shoe_model').deleteOne({_id:new ObjectId(id)});
-        const cursorFind = client.db('inventory').collection('shoe_model').find();
-        cursorFind.sort({"model":1});
-        const results = await cursorFind.toArray();
-        if(results.length)res.status(200).send({message: results});
-        emptyData();
-        refreshData();
-        tempDataForGet=undefined;
+        const result=req.body.pw;
+        if(result==='arsenal') {
+            await client.connect();
+            const result = await client.db('inventory').collection('shoe_model').deleteOne({_id:new ObjectId(id)});
+            const cursorFind = client.db('inventory').collection('shoe_model').find();
+            cursorFind.sort({"model":1});
+            const results = await cursorFind.toArray();
+            if(results.length)res.status(200).send({message: results});
+            emptyData();
+            refreshData();
+            tempDataForGet=undefined;
+        } else {
+            res.status(500).send({message: 'Wrong password'});
+        }
     } catch(e) { 
         console.error('e in shoes delete: ', e);
         res.status(500).send({message: e});
